+++ tests/testthat/test-model_tests.R
@@ -1,5 +1,6 @@
 # # Making sure the tests family of functions works as intended.
-marvel_friends <- manynet::to_giant(manynet::to_unsigned(manynet::ison_marvel_relationships)) %>%
+marvel_friends <- manynet::to_uniplex(manynet::fict_marvel, "relationship") %>% 
+  manynet::to_giant() %>% manynet::to_unsigned() %>% 
   manynet::to_subgraph(PowerOrigin == "Human")
 cugtest <- test_random(marvel_friends,
                        manynet::net_heterophily,
@@ -38,9 +39,6 @@ test_that("test_random works", {
 })
 
 # Set the qaptest up
-marvel_friends <- manynet::to_unsigned(manynet::ison_marvel_relationships)
-marvel_friends <- manynet::to_giant(marvel_friends)
-marvel_friends <- manynet::to_subgraph(marvel_friends, PowerOrigin == "Human")
 qaptest <- test_permutation(marvel_friends,
                             manynet::net_heterophily,
                             attribute = "Attractive",