+++ NEWS.md
@@ -1,3 +1,11 @@
+# migraph 1.5.7
+
+2026-02-18
+
+- Using fict_marvel in tests
+- Added tidy(), glance(), and summary() methods for sienaFit class objects
+- Added summary.ergm()
+
 # migraph 1.5.6
 
 2025-11-19